refactor: 拆分agent快照窗口组件hooks
